Moved object table
