+++ docs/diagrams/save/smcore.drawio
@@ -1,6 +1,6 @@
-<mxfile host="Electron" modified="2024-10-03T20:44:22.608Z" agent="Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/22.1.2 Chrome/114.0.5735.289 Electron/25.9.4 Safari/537.36" etag="8IMDY4YhSqVKrtehgwen" version="22.1.2" type="device">
+<mxfile host="Electron" modified="2024-10-03T21:08:42.100Z" agent="Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/22.1.2 Chrome/114.0.5735.289 Electron/25.9.4 Safari/537.36" etag="FoBZx1cLNXYHDAgy4dfG" version="22.1.2" type="device">
   <diagram name="Page-1" id="VDByYd5_aM_RI3PdM5R2">
-    <mxGraphModel dx="1418" dy="868" grid="0" gridSize="10" guides="1" tooltips="1" connect="1" arrows="1" fold="1" page="1" pageScale="1" pageWidth="1920" pageHeight="1200" math="0" shadow="0">
+    <mxGraphModel dx="1418" dy="868" grid="1" gridSize="10" guides="1" tooltips="1" connect="1" arrows="1" fold="1" page="1" pageScale="1" pageWidth="1920" pageHeight="1200" math="0" shadow="0">
       <root>
         <mxCell id="0" />
         <mxCell id="1" parent="0" />
