+++ simula24/objectmanager/CMakeLists.txt
@@ -1,5 +1,6 @@
 
 set(OBJECT_MANAGER_SOURCES
-    ${CMAKE_CURRENT_SOURCE_DIR}/object/ObjectTable.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/ObjectTable.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/ObjectManager.cpp
     PARENT_SCOPE
 )
