+++ simula24/objectmanager/ObjectTable.cpp
