+++ simula24/objectmanager/ObjectTable.h
