+++ simula24/smcore/Simulation.h
@@ -1,7 +1,9 @@
 #ifndef SMCORE_SIMULATION_H_
 #define SMCORE_SIMULATION_H_
 
-
+#include <core/stl/array.h>
+#include <smcore/entity/Civilian.h>
+#include <smcore/map/ObjectMap.h>
 namespace simula24
 {
 
@@ -10,8 +12,8 @@ class Simulation
 public:
 
 private:
-
-
+    stl::array<Civilian> m_civlist;
+    ObjectMap m_objectMap;
 };
 
 }
